requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
llama-cpp-python>=0.2.0
numpy>=1.21.0
aiofiles>=0.8.0
python-multipart>=0.0.5
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
